[refactor] projectのマイグレーションを関数切り出し (#1967)
* projectのschemaをdomainディレクトリ以下に移動

* 移動し忘れ

* noteSchema移動し忘れ

* たぶんこう

* CharacterInfo差し戻し

* anyは仕方ない気がする

* import ミス

* ミス

* ProjectFileFormatError型を用意

* causeいらない
